requirements.txt
django
djangorestframework
django-cors-headers
python-decouple
djangorestframework-simplejwt